requirements.txt
streamlit
pandas
google-generativeai
openpyxl
pypdf
python-dotenv
openai
tenacity
